+++ src/server/modules/catalog/catalog.service.ts
@@ -419,7 +419,6 @@ export class CatalogService {
             "themeFont", "themeBlur", "themeOverlayOpacity", "communityLink", "chatEnabled", "boardEnabled", "siteHandle",
             "hideLive", "hideStore", "hideSocial", "hideNetwork", "hideDig", "hideSamples", "hideCollab", "hideLab",
             "membershipMonthlyPrice", "peerEnabled", "peerAllowDownloads", "peerFederation",
-            "peerChatEnabled", "peerChatGuestEnabled",
             "brandPrimary", "brandAccent", "instanceProfile",
             "adminTreasuryAddress", "walletAddress", "adminFeePercentage", "web3Enabled",
             "allowPublicRegistration", "listenerSelfPublish", "publicUrl"
@@ -448,8 +447,6 @@ export class CatalogService {
                     else if (k === 'peerEnabled') res[k] = 'false';
                     else if (k === 'peerAllowDownloads') res[k] = 'true';
                     else if (k === 'peerFederation') res[k] = 'false';
-                    else if (k === 'peerChatEnabled') res[k] = 'false';
-                    else if (k === 'peerChatGuestEnabled') res[k] = 'false';
                     else res[k] = '';
                 }
             }